ocr_backend/lambda/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai==1.3.0
python-dotenv==1.0.0
Pillow==10.1.0
pytesseract==0.3.10
mangum==0.17.0
reportlab==4.0.9
httpx==0.25.2
boto3==1.34.69
